Fix Vimeo providers’ `available?` method w/ password-protected videos
It seems like one point in the past, Vimeo's password-protected videos contained an element with the class `exception_title--password`. That is no longer the case. Moreover, password-protected videos return a 200 status code, so the base provider class' `available?` method also doesn't work.

For the Vimeo scraper provider, we look for json-ld. If none is present, we assume the video is not available.

For the Vimeo API provider, we check if we can get a basic property from the Vimeo API for the video, such as the name.