Get author portrait from JSON-LD since old method fails
The Vimeo Scraper provider's old `author_thumbnail` method looks for an image source in an inline script. This was failing with more recent Vimeo pages. Luckily the JSON-LD contains the author portrait.

Unfortunately the JSON-LD reports a different image size than the Vimeo API. So the `author_thumbnail` test will need to be updated to account for that.