Remove redundant test
This test seems to duplicate what the `with video 136971428` context's `stats` test does. Additionally it seems like it is expecting an impossilbe result, unless the `7848846` video somehow really has no plays, likes, or comments.

More the point, for whatever reason, I'm unable to re-record this cassette because the scraper is running into Cloudflare's Turnstile. So to get the tests passing, I'm thinking maybe we can do without this test?